GUI changes
Renamed variables in myDateTimeModel_propertyChange()

Added functionality to Auto StartTime checkbox, which is also triggered by the Do All button.  OScontrolsPanel had to have the checkbox copied in as a member so it could be accessed.

For NGPS-PLAN, removed the Fetch Live checkbox and disabled the Connections window menu item.

Tweaked parsing of OTMflags.  Removed WARN-OTM state.  Targets stay PENDING if OTMflag is non-empty but switch to ERROR-OTM if we see "DAY" and "-1".

Renamed fileMenu variable.

Added units to Seeing label.

Added "Sun and Moon" menu item to pop-up a window with rise/set times.

Changed column headers in main table.